created DarkSkyNetworkManager to facilitate DarkSkyAPI related calls. DarkSkyNetworkManager holds the baseURL and api secret key so any DarkSky related call can be sent through that manager + updated WeatherDataService to use the DarkSkyNetworkManager to make its forecast call
