+++ JennyWeather/Network/DataService/WeatherDataService.swift
@@ -15,12 +15,7 @@ protocol WeatherDataServiceGettable {
 	func getWeatherData(latitude: Double, longitude: Double, success: @escaping WeatherDataServiceSuccess, failure: DataServiceFailure)
 }
 
-class WeatherDataService: NSObject, URLSessionDataDelegate {
-	
-	private static let secretKey = "9e81c40beac3dc5695160dec5db8258b"
-	private static let url = "https://api.darksky.net/forecast/\(secretKey)/"
-	private var urlSession: URLSession?
-	private var observation: NSKeyValueObservation?
+class WeatherDataService: NSObject {
 	
 	private static let unitsParameterName = "units"
 	private static var unitsParameterValue: String {
@@ -28,36 +23,22 @@ class WeatherDataService: NSObject, URLSessionDataDelegate {
 		return units
 	}
 	
-	override init() {
-		super.init()
-
-		let configuration = URLSessionConfiguration.default
-		let delegate = self
-		urlSession = URLSession(configuration: configuration, delegate: delegate, delegateQueue: nil)
-	}
-	
 }
 
 extension WeatherDataService: WeatherDataServiceGettable {
 	
-	private func getURL(latitude: Double, longitude: Double) -> URL? {
-		let urlString = WeatherDataService.url + "\(latitude),\(longitude)"
+	private func getURLString(latitude: Double, longitude: Double) -> String {
+		let urlString = "\(latitude),\(longitude)"
 		
 		// TODO: If need more parameters, build a more robust way or use Alamofire
 		let parameterString = "?" + "\(WeatherDataService.unitsParameterName)=\(WeatherDataService.unitsParameterValue)"
 		let newURLString = urlString + parameterString
-		
-		let url = URL(string: newURLString)
-		return url
+		return newURLString
 	}
 	func getWeatherData(latitude: Double, longitude: Double, success: @escaping WeatherDataServiceSuccess, failure: DataServiceFailure) {
-		guard let url = getURL(latitude: latitude, longitude: longitude) else {
-			failure?(nil)
-			return
-		}
-		let request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalAndRemoteCacheData, timeoutInterval: 10)
+		let urlString = getURLString(latitude: latitude, longitude: longitude)
 		
-		let dataTask = URLSession.shared.dataTask(with: request) { (data, response, error) in
+		DarkSkyNetworkManager.shared.sendForecastRequest(urlString: urlString) { (data, response, error) in
 			if let sureData = data,
 				let json = try? JSONSerialization.jsonObject(with: sureData, options: []),
 				let jsonDict = json as? [String: Any] {
@@ -77,11 +58,6 @@ extension WeatherDataService: WeatherDataServiceGettable {
 				failure?(error)
 			}
 		}
-		
-		observation = dataTask.progress.observe(\.fractionCompleted) { progress, _ in
-		  print("progress: ", progress.fractionCompleted)
-		}
-		dataTask.resume()
 	}
 }
